requirements.txt
django>=2.1
psycopg2
djangorestframework==3.8.2
requests